move SnackbarContext to ui project, rename, export
