+++ platform/ui/src/components/snackbar/SnackbarContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import SnackbarItem from './SnackbarItem';
 import './Snackbar.css';
-import { useSnackbarContext } from '../../../../viewer/src/context/SnackbarContext';
+import { useSnackbarContext } from '../../utils/SnackbarProvider';
 
 const SnackbarContainer = () => {
   const { snackbarItems, hide } = useSnackbarContext();
